Cleanup Zephyr and remove redundant platform.h
