+++ apps/microtvm/zephyr/template_project/CMakeLists.txt.template
@@ -54,13 +54,6 @@ else()
   set(cmsis_includes "")
 endif()
 
-# define a library for the model sources.
-zephyr_library_named(tvm_model)
-file(GLOB_RECURSE tvm_model_srcs model/codegen/host/src/*.c model/codegen/host/lib/*.o)
-target_sources(tvm_model PRIVATE ${tvm_model_srcs})
-target_include_directories(tvm_model PRIVATE ${CMAKE_SOURCE_DIR}/include crt_config crt/include ${cmsis_includes})
-target_compile_options(tvm_model PRIVATE -Wno-unused-variable)  # TVM-generated code tends to include lots of these.
-
 set(CRT_LIBS <API_SERVER_CRT_LIBS>)
 set(CRT_LIB_BASE crt/src/runtime/crt)
 foreach(crt_lib_name ${CRT_LIBS})
@@ -68,13 +61,15 @@ foreach(crt_lib_name ${CRT_LIBS})
   file(GLOB_RECURSE crt_lib_srcs ${CRT_LIB_BASE}/${crt_lib_name}/*.c ${CRT_LIB_BASE}/${crt_lib_name}/*.cc)
   target_sources(${crt_lib_name} PRIVATE ${crt_lib_srcs})
   zephyr_library_include_directories(${crt_lib_name} PRIVATE crt_config crt/include)
-  
-  # Link CRT lib to app
   target_link_libraries(app PRIVATE ${crt_lib_name})
-  # Link CRT lib to tvm_model to use CRT references
-  target_link_libraries(tvm_model PRIVATE ${crt_lib_name})
 endforeach(crt_lib_name ${CRT_LIBS})
 
+# define a library for the model sources.
+zephyr_library_named(tvm_model)
+file(GLOB_RECURSE tvm_model_srcs model/codegen/host/src/*.c model/codegen/host/lib/*.o)
+target_sources(tvm_model PRIVATE ${tvm_model_srcs})
+target_include_directories(tvm_model PRIVATE ${CMAKE_SOURCE_DIR}/include crt_config crt/include ${cmsis_includes})
+target_compile_options(tvm_model PRIVATE -Wno-unused-variable)  # TVM-generated code tends to include lots of these.
 target_link_libraries(app PRIVATE tvm_model)
 
 if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/src/api)
