+++ apps/microtvm/zephyr/template_project/src/host_driven/platform.c
@@ -17,28 +17,48 @@
  * under the License.
  */
 
-#include "tvm/platform.h"
+/*!
+ * \brief Implementation of TVMPlatform functions in tvm/runtime/crt/platform.h
+ */
 
+#include <dlpack/dlpack.h>
 #include <stdarg.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <tvm/runtime/crt/error_codes.h>
+#include <zephyr/drivers/gpio.h>
 #include <zephyr/kernel.h>
+#include <zephyr/random/rand32.h>
+#include <zephyr/sys/printk.h>
 #include <zephyr/sys/reboot.h>
+#include <zephyr/timing/timing.h>
 
-#include "dlpack/dlpack.h"
-#include "tvm/runtime/crt/error_codes.h"
-
-// Heap for use by TVMPlatformMemoryAllocate.
 K_HEAP_DEFINE(tvm_heap, HEAP_SIZE_BYTES);
 
-// Called by TVM when a message needs to be formatted.
+volatile timing_t g_microtvm_start_time, g_microtvm_end_time;
+int g_microtvm_timer_running = 0;
+
+#ifdef CONFIG_LED
+#define LED0_NODE DT_ALIAS(led0)
+static const struct gpio_dt_spec led0 = GPIO_DT_SPEC_GET(LED0_NODE, gpios);
+#endif  // CONFIG_LED
+
+// This is invoked by Zephyr from an exception handler, which will be invoked
+// if the device crashes. Here, we turn on the LED and spin.
+void k_sys_fatal_error_handler(unsigned int reason, const z_arch_esf_t* esf) {
+#ifdef CONFIG_LED
+  gpio_pin_set_dt(&led0, 1);
+#endif
+  for (;;)
+    ;
+}
+
 __attribute__((weak)) size_t TVMPlatformFormatMessage(char* out_buf, size_t out_buf_size_bytes,
                                                       const char* fmt, va_list args) {
   return vsnprintk(out_buf, out_buf_size_bytes, fmt, args);
 }
 
-// Called by TVM when an internal invariant is violated, and execution cannot continue.
 __attribute__((weak)) void TVMPlatformAbort(tvm_crt_error_t error) {
   TVMLogf("TVMError: 0x%x", error);
   sys_reboot(SYS_REBOOT_COLD);
@@ -49,15 +69,81 @@ __attribute__((weak)) void TVMPlatformAbort(tvm_crt_error_t error) {
     ;
 }
 
-// Called by TVM when memory allocation is required.
 __attribute__((weak)) tvm_crt_error_t TVMPlatformMemoryAllocate(size_t num_bytes, DLDevice dev,
                                                                 void** out_ptr) {
   *out_ptr = k_heap_alloc(&tvm_heap, num_bytes, K_NO_WAIT);
   return (*out_ptr == NULL) ? kTvmErrorPlatformNoMemory : kTvmErrorNoError;
 }
 
-// Called by TVM to free an allocated memory.
 __attribute__((weak)) tvm_crt_error_t TVMPlatformMemoryFree(void* ptr, DLDevice dev) {
   k_heap_free(&tvm_heap, ptr);
+  return kTvmErrorNoError;
+}
+
+// Called to start system timer.
+__attribute__((weak)) tvm_crt_error_t TVMPlatformTimerStart() {
+  if (g_microtvm_timer_running) {
+    TVMLogf("timer already running");
+    return kTvmErrorPlatformTimerBadState;
+  }
+
+#ifdef CONFIG_LED
+  gpio_pin_set_dt(&led0, 1);
+#endif
+  g_microtvm_start_time = timing_counter_get();
+  g_microtvm_timer_running = 1;
+  return kTvmErrorNoError;
+}
+
+// Called to stop system timer.
+__attribute__((weak)) tvm_crt_error_t TVMPlatformTimerStop(double* elapsed_time_seconds) {
+  if (!g_microtvm_timer_running) {
+    TVMLogf("timer not running");
+    return kTvmErrorSystemErrorMask | 2;
+  }
+
+#ifdef CONFIG_LED
+  gpio_pin_set_dt(&led0, 0);
+#endif
+
+  g_microtvm_end_time = timing_counter_get();
+  uint64_t cycles = timing_cycles_get(&g_microtvm_start_time, &g_microtvm_end_time);
+  uint64_t ns_spent = timing_cycles_to_ns(cycles);
+  *elapsed_time_seconds = ns_spent / (double)1e9;
+  g_microtvm_timer_running = 0;
+  return kTvmErrorNoError;
+}
+
+__attribute__((weak)) tvm_crt_error_t TVMPlatformGenerateRandom(uint8_t* buffer, size_t num_bytes) {
+  uint32_t random;  // one unit of random data.
+
+  // Fill parts of `buffer` which are as large as `random`.
+  size_t num_full_blocks = num_bytes / sizeof(random);
+  for (int i = 0; i < num_full_blocks; ++i) {
+    random = sys_rand32_get();
+    memcpy(&buffer[i * sizeof(random)], &random, sizeof(random));
+  }
+
+  // Fill any leftover tail which is smaller than `random`.
+  size_t num_tail_bytes = num_bytes % sizeof(random);
+  if (num_tail_bytes > 0) {
+    random = sys_rand32_get();
+    memcpy(&buffer[num_bytes - num_tail_bytes], &random, num_tail_bytes);
+  }
+  return kTvmErrorNoError;
+}
+
+__attribute__((weak)) tvm_crt_error_t TVMPlatformInitialize() {
+#ifdef CONFIG_LED
+  if (!device_is_ready(led0.port)) {
+    for (;;)
+      ;
+  }
+  int ret = gpio_pin_configure_dt(&led0, GPIO_OUTPUT_ACTIVE);
+  if (ret < 0) {
+    TVMPlatformAbort((tvm_crt_error_t)0xbeef4);
+  }
+  gpio_pin_set_dt(&led0, 0);
+#endif
   return kTvmErrorNoError;
 }
